Add files via upload
This is just because i needed one and couldn't find any, the code is very basic and there is a lot of things that could have been done better. The header and footer is completely optional and purely just to keep all the <head> mess out of the code. Index is the Calendar, Databox is the Iframe Content, nothing needs to be edited to run